Create subs line item in a helper module
This module allows the behaviour of creating subscription line items
after an OrderContents#add to be shared between the spree controllers
that need it.

It requires the modified line item to be passed to it since the instance
variables for the spree front end controllers are different than the api
controllers
